Support namespace average_node_cpu_frequency in the SDK
** Problem **

In order to support consistent scaling for containers running in a cluster where nodes have a mix of different CPU speeds, we must be able to get the quota commodities that are in MHz in the market back into millicore units.

** Solution **

Kubeturbo converts namespace quota capacities from millicores to MHz by multiplying the millicore values by the average of all node CPU frequencies in the entire kubernetes cluster. By providing this value as a parameter on the Namespace, we can invert this conversion where necessary to go back to millicores from MHz.

** Testing Done **

* Verified in-app that we are able to send average_node_cpu_frequency end-to-end from kubeturbo out through the TP broadcast.
* New unit tests.